+++ packages/components/src/templates/next/components/internal/SearchableTable/Native/NativeSearchableTable.tsx
@@ -1,6 +1,6 @@
 import type { NativeSearchableTableProps } from "~/interfaces"
 import { SearchableTableClient } from "../shared"
-import { HYPERLINK_EXCEL_FUNCTION } from "./constants"
+import { HYPERLINK_EXCEL_FUNCTION } from "../shared/constants"
 
 export const NativeSearchableTable = ({
   items,
