+++ packages/components/src/templates/next/components/internal/SearchableTable/shared/CellContent.tsx
@@ -1,7 +1,7 @@
 import type { IsomerSiteProps, LinkComponentType } from "~/types"
 import { getReferenceLinkHref } from "~/utils"
 import BaseParagraph from "../../BaseParagraph"
-import { HYPERLINK_EXCEL_FUNCTION } from "../Native/constants"
+import { HYPERLINK_EXCEL_FUNCTION } from "./constants"
 
 interface CellContentProps {
   content: string | number
