+++ packages/components/src/templates/next/components/internal/SearchableTable/shared/constants.ts
